backend/requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
flask-cors==4.0.0
